requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
# ── Free-tier provider chain (cascading fallback) ──────────────────────────
# AI chain:    anthropic → gemini → groq → openrouter → cerebras → together → mistral → mock
# Scout chain: perplexity → tavily → duckduckgo → exa → ai_scout
# Switch to production: set billing + reorder chain
google-genai==1.70.0
tavily-python==0.7.23
groq>=0.11.0
duckduckgo-search>=6.0.0
# Cerebras, Together AI, Mistral use httpx (already in requirements — no extra SDK needed)
anyio==4.12.1
attrs==25.4.0
cachetools==6.2.6
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
cryptography==46.0.5
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
fastapi==0.135.1
frozenlist==1.8.0
fsspec==2026.2.0
google-api-core==2.30.0
google-auth==2.49.1
google-cloud-run==0.15.0
google-cloud-secret-manager==2.26.0
googleapis-common-protos==1.73.0
grpc-google-iam-v1==0.14.3
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
iniconfig==2.3.0
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.19
langgraph==1.1.2
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.11
langsmith==0.7.17
markdown-it-py==4.0.0
mdurl==0.1.2
mmh3==5.2.1
multidict==6.7.1
neo4j==6.1.0
orjson==3.11.7
ormsgpack==1.12.2
packaging==26.0
pluggy==1.6.0
postgrest==2.28.2
propcache==0.4.1
proto-plus==1.27.1
protobuf==6.33.5
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
PyGithub==2.8.1
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
PyNaCl==1.6.2
pyparsing==3.3.2
pyroaring==1.0.3
pytest==9.0.2
pytest-asyncio==1.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-telegram-bot==22.6
pytz==2026.1.post1
PyYAML==6.0.3
realtime==2.28.2
requests==2.32.5
requests-toolbelt==1.0.0
rich==14.3.3
six==1.17.0
sniffio==1.3.1
starlette==0.52.1
storage3==2.28.2
StrEnum==0.4.15
strictyaml==1.7.3
structlog==25.5.0
supabase==2.28.2
supabase-auth==2.28.2
supabase-functions==2.28.2
tenacity==9.1.4
tornado==6.5.5
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uv==0.10.10
uvicorn==0.41.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
# Image processing (design engine)
Pillow>=10.2.0,<11.0
